coreboot: Add support for FMAP and multiple CBFS
ROM images with a FMAP may have multiple CBFS.
Scan all available CBFS so that, say, a SeaBIOS bootable image doesn't
have to be in the main CBFS.

Coreboot puts the FMAP location in the BOOT_MEDIA_PARAMS entry in the
coreboot table. We can grab that and can all regions for CBFS files.

Signed-off-by: Ben Gardner <gardner.ben@gmail.com>